# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/maritime/islands/{id}": {
      "get": {
        "tags": [
          "涉海档案 · 海岛"
        ],
        "summary": "查询海岛完整档案",
        "operationId": "detail_7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIslandView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "涉海档案 · 海岛"
        ],
        "summary": "编辑或启停海岛，需最新版本",
        "operationId": "update_9",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteIsland"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIslandView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "涉海档案 · 海岛"
        ],
        "summary": "删除未被引用的海岛，需最新版本",
        "operationId": "delete_7",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/maritime/islands": {
      "get": {
        "tags": [
          "涉海档案 · 海岛"
        ],
        "summary": "分页查询海岛",
        "operationId": "list_9",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inhabitationType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResultIslandView"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "涉海档案 · 海岛"
        ],
        "summary": "新建海岛",
        "operationId": "create_10",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteIsland"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIslandView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
...
      "WriteIsland": {
        "type": "object",
        "properties": {
          "version": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "新建不传；编辑必填，缺失返回400，旧版本返回409",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "region": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "location": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "minLength": 1,
            "pattern": "ENABLED|DISABLED"
          },
          "inhabitationType": {
            "type": "string",
            "minLength": 1,
            "pattern": "INHABITED|UNINHABITED"
          }
        },
        "required": [
          "inhabitationType",
          "name",
          "region",
          "status"
        ]
      },
      "ApiResponseIslandView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/IslandView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
      "IslandView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "inhabitationType": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "fixtureKey": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "createdAt",
          "fixtureKey",
          "id",
          "inhabitationType",
          "location",
          "name",
          "region",
          "status",
          "updatedAt",
          "version"
        ]
      },
...
      "ApiResponsePageResultIslandView": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResultIslandView"
          },
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      },
      "PageResultIslandView": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IslandView"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "requestId": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "description": "字段级校验错误；无字段错误时为 null",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        },
        "required": [
          "code",
          "data",
          "message",
          "requestId"
        ]
      }
    }
  }
}